aider: Added an `equals` method and a `hashCode` method to the `Item` class to compare items by their name in the `Inventory` class.
# Aider chat conversation:

USER: Wow, amazing! Thank you!

ASSISTANT: You're welcome! If you have any other requests or need further modifications, feel free to ask. I'm here to help!

USER: I have added inventory item instances and added these to the rooms. However, when I play the game, I can successfully pick up items using a command like GET COMPASS. But the DROP command is not working. If I enter DROP COMPASS, it says "You don't have that item" even though it is in my inventory.

ASSISTANT: The issue is likely due to how the `Item` objects are being compared in the `Inventory` class. In Java, when you use the `contains` or `remove` method with an object, it uses the `equals` method to compare the objects. Since the `Item` class does not override the `equals` method, it uses the default `equals` method from the `Object` class, which compares the memory addresses of the objects, not their contents.

To fix this, we need to override the `equals` method in the `Item` class to compare the `name` of the items. This way, two `Item` objects with the same `name` will be considered equal.

Here is the change we need to make:

```java
Test1\TextAdventure.java
<<<<<<< HEAD
    static class Item {
        String name;

        Item(String name) {
            this.name = name;
        }

        public String toString() {
            return name;
        }
    }
=======
    static class Item {
        String name;

        Item(String name) {
            this.name = name;
        }

        public String toString() {
            return name;
        }

        @Override
        public boolean equals(Object obj) {
            if (this == obj) {
                return true;
            }
            if (obj == null || getClass() != obj.getClass()) {
                return false;
            }
            Item item = (Item) obj;
            return Objects.equals(name, item.name);
        }

        @Override
        public int hashCode() {
            return Objects.hash(name);
        }
    }
>>>>>>> updated
```

This change adds an `equals` method that compares the `name` of the `Item` objects, and a `hashCode` method that returns a hash code value for the object, which is required when you override `equals`.