aider: Added migration file to create the `TodoItems` table in the database.
# Aider chat conversation:

USER: This is looking good. Now to display to-do items, we'll need to be able to add them, but before we get there, I see we are using Entity Framework. Do we need to create any migration files to create the database schema?

ASSISTANT: Yes, you're correct. When using Entity Framework Core, we need to create a migration that will generate the SQL statements necessary to create the database schema.

To create a migration, we typically use the `dotnet ef migrations add` command in the terminal. However, since we're working in this chat-based environment, I'll provide you with the code that would be generated by this command.

The migration file will be added to a directory named `Migrations` in the root of your project. The file will have a name like `yyyyMMddHHmmss_InitialCreate.cs`, where `yyyyMMddHHmmss` is a timestamp indicating when the migration was created.

This file will contain a `Up` method, which defines the changes to apply to the database (i.e., creating the `TodoItems` table), and a `Down` method, which defines how to revert these changes.

Please add the `Migrations` directory to the chat so I can provide the migration file.

USER: add Test2/Migrations

ASSISTANT: Great! Now that we have the `Migrations` directory, we can create the migration file.

The migration file will contain the necessary code to create the `TodoItems` table in the database. The `Up` method will define the creation of the table, and the `Down` method will define the deletion of the table.

Here's the migration file:

```csharp
Test2/Migrations/20220404123456_InitialCreate.cs
<<<<<<< HEAD

=======
using Microsoft.EntityFrameworkCore.Migrations;

namespace TodoList.Migrations
{
    public partial class InitialCreate : Migration
    {
        protected override void Up(MigrationBuilder migrationBuilder)
        {
            migrationBuilder.CreateTable(
                name: "TodoItems",
                columns: table => new
                {
                    Id = table.Column<int>(nullable: false)
                        .Annotation("SqlServer:Identity", "1, 1"),
                    Task = table.Column<string>(nullable: false),
                    IsComplete = table.Column<bool>(nullable: false)
                },
                constraints: table =>
                {
                    table.PrimaryKey("PK_TodoItems", x => x.Id);
                });
        }

        protected override void Down(MigrationBuilder migrationBuilder)
        {
            migrationBuilder.DropTable(
                name: "TodoItems");
        }
    }
}
>>>>>>> updated
```

This migration file will create a table named `TodoItems` with columns `Id`, `Task`, and `IsComplete`. The `Id` column is the primary key. The `Task` column is required, and the `IsComplete` column is a boolean indicating whether the task is complete.